edit database

+++ app/models/receipt.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Enum
+from sqlalchemy import Column, Integer, String, ForeignKey, Enum , Numeric
 from sqlalchemy.orm import relationship
 from app.database import Base
 import enum
@@ -17,7 +17,7 @@ class Receipt(Base):
     id                  = Column(Integer, primary_key=True, index=True)
     user_id             = Column(Integer, ForeignKey("users.id"),            nullable=False)
     shipping_address_id = Column(Integer, ForeignKey("shipping_address.id"), nullable=False)
-    total_price         = Column(Integer, nullable=False)
+    total_price         = Column(Numeric(10,2), nullable=False)
     payment_status      = Column(String(50),    nullable=False)
     status              = Column(Enum(ReceiptStatus), nullable=False, default=ReceiptStatus.PENDING)
